+++ src/test/java/StatsServiceTest.java
@@ -3,7 +3,7 @@ import org.junit.jupiter.api.Test;
 import ru.netology.stats.StatsService;
 
 public class StatsServiceTest {
-    int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+    long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
     @Test
     public void sumAllSalesTest() {
